Things ARE happening
There are rectangles (Tiles) and they are rotating at the top of the
screen, however they are not yet falling.
Also the BackKey is now being recognized but only changes the background
color to red since i can't get it to bring up the menu screen back up.
Able to detect touches on the screen and get their X and Y coordinates
but right now touching the screen only changes the background color
between blue and green.
.... it's progress i guess.